+++ crates/fluvio-benchmark/src/bin/fbm.rs
@@ -67,7 +67,7 @@ fn main() {
                 run_block_on(
                     all_stats
                         .lock()
-                        .map(|a| println!("{}", a.compare_stats(&config, &other))),
+                        .map(|a| println!("{}", a.compare_stats(&config, other))),
                 );
             }
         }
@@ -83,7 +83,7 @@ fn main() {
 
 async fn take_stats(all_stats: AllStatsSync) -> AllStats {
     let mut guard = all_stats.lock().await;
-    mem::replace(&mut *guard, AllStats::default())
+    mem::take(&mut *guard)
 }
 fn benchmarking_dir() -> Result<PathBuf, BenchmarkError> {
     let dir_path = fluvio_base_dir()?.join("benchmarks");
